+++ athena-scripts/raw_stage_optimisation_solution/scripts/etl/strategies/Strategy.py
@@ -5,10 +5,10 @@ from datetime import datetime
 
 import pandas as pd
 
-from ..old.core_preprocessing_functions import (add_new_column, add_new_column_from_struct, empty_string_to_null,
-                                                generate_key_value_records, remove_columns, remove_row_duplicates,
-                                                remove_rows_missing_mandatory_values, rename_column_names)
-from ..util.processing_utilities import extract_element_by_name_and_validate
+from ..util.processing_utilities import (add_new_column, add_new_column_from_struct, empty_string_to_null,
+                                         extract_element_by_name_and_validate, generate_key_value_records,
+                                         remove_columns, remove_row_duplicates, remove_rows_missing_mandatory_values,
+                                         rename_column_names)
 
 
 class Strategy:
