+++ athena-scripts/raw_stage_optimisation_solution/scripts/etl/util/exceptions/TableQueryException.py
@@ -1,0 +1,4 @@
+class TableQueryException(Exception):
+    """
+    Custom exception class for representing errors querying table in Database.
+    """
